write to tmp.

+++ sge_submit.sh
@@ -16,9 +16,11 @@ for Subject in $(ls -d *); do
 	#fi	
 
 	for r in $(seq 1 1 20); do
-		sed "s/s in 106/s in ${Subject}/g; s/run1/run${r}/g " < ${SCRIPT}/preproc_functional.sh > ${SCRIPT}/tmp/functional_${Subject}_${r}.sh 
-		qsub -V -M kaihwang -m e -e ~/tmp -o ~/tmp ${SCRIPT}/tmp/functional_${Subject}_${r}.sh   #-l mem_free=5G
-
+		
+		if [ ! -e "${WD}/${Subject}/run${r}/nswdkmt_run${r}_raw_6.nii.gz" ]; then
+			sed "s/s in 106/s in ${Subject}/g; s/run1/run${r}/g " < ${SCRIPT}/preproc_functional.sh > ${SCRIPT}/tmp/functional_${Subject}_${r}.sh 
+			qsub -V -M kaihwang -m e -e ~/tmp -o ~/tmp ${SCRIPT}/tmp/functional_${Subject}_${r}.sh   #-l mem_free=5G
+		fi
 	done
 
 done
